Port reduce_over_samples tests to pytest
